requirements.txt
PySide6>=6.6.0
requests>=2.31.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
cryptography>=40.0.0
pandas>=2.0.0
openpyxl>=3.1.0
notion-client>=2.2.0 